Refactor the DDSGapproximation routine
