--- src/global/simulation.jl
@@ -1,47 +1,0 @@
-using AxisArrays: AxisArray
-function simulate!(context, grid, periods, policy_variables, state_variables, replications=1)
-    model = context.models[1]
-    params = context.work.params
-    results = context.results.model_results[1]
-    trends = results.trends
-    perfect_foresight_ws = PerfectForesightWs(context, periods)
-    shocks = perfect_foresight_ws.shocks
-    initial_values = get_dynamic_initialvalues(context)
-    endogenous_nbr = model.endogenous_nbr
-    exogenous_nbr = model.exogenous_nbr
-    steadystate = trends.endogenous_steady_state
-    Sigma_e = model.Sigma_e
-    chol_sigma_e_L = cholesky(Sigma_e).L
-    x = reshape(perfect_foresight_ws.x, exogenous_nbr, periods)
-        display(x)
-    
-    Y = Array{Float64}(undef, periods, endogenous_nbr + exogenous_nbr, replications)
-    y = Vector{Float64}(undef, 3*endogenous_nbr)
-    z = Vector{Float64}(undef, endogenous_nbr)
-    random_shocks = Matrix{Float64}(undef, exogenous_nbr, periods)
-    sv_buffer = Vector{Float64}(undef, length(state_variables))
-    pv_buffer = Vector{Float64}(undef, length(policy_variables))
-    @inbounds for r in 1:replications
-        mul!(random_shocks, chol_sigma_e_L, randn(exogenous_nbr, periods))
-        if !isempty(shocks)
-            random_shocks .+= x
-        end 
-        y[1:endogenous_nbr] .= initial_values
-        for p in 1:periods
-            @views begin
-                preamble_block(y, random_shocks[:, p], params, steadystate)
-                sv_buffer .= y[state_variables]
-                evaluateBatch!(pv_buffer, grid, sv_buffer)
-                y[policy_variables] .= pv_buffer
-                Y[p, 1:endogenous_nbr, r] .= y[endogenous_nbr .+ (1:endogenous_nbr)]
-                Y[p, endogenous_nbr .+ (1:exogenous_nbr), r] .= random_shocks[:, p]
-
-            end  
-            circshift!(y, -endogenous_nbr)
-        end
-    end
-    symboltable = context.symboltable
-    varnames = vcat(Symbol.(get_endogenous(symboltable)),
-                    Symbol.(get_exogenous(symboltable)))
-    return AxisArray(Y, Undated(1):Undated(periods), varnames, 1:replications) 
-end
